Assume users have only one device of a type
That is, a user only has at max one iPhone. This assumption is
obviously false, but this assumption is necessary due to the way the
database is structured. In order to change this, you can specify
device ID's in the Call model. However, web clients do not have
"device ID"s, so a better solution would be to make those fields into
the ID of a device. This means that we'll have to make device into its
own custom type.